requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiohttp==3.9.1
python-dotenv==1.0.0
apscheduler==3.10.4
pytz==2023.3
openai==1.3.5
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.100.0
python-multipart==0.0.6
